requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic[email]==2.10.0
sqlalchemy==2.0.36
pymysql==1.1.0
bcrypt==4.2.1
python-multipart==0.0.6
pyjwt==2.11.0
python-dotenv==1.0.0
